close connection on idle
I found that long-standing connections sometimes enter a state where
the error reports `SqliteError { code: 11, message: "database disk
image is malformed" }`. And restarting the service fixes the problem.

I suspect this could be caused by certain page operations in the
sqlite file that somewhat modifies the sqlite file and WAL. Because
wiki-search opens the database immutably (assuming fs is readonly), it
cannot detect the needed change.

This commit makes the connection close after 5 minutes of idle and
re-opens it as needed with a connection pool.